Fix io/h5 build for aarch64
Include and library paths for HDF5 are architecture-specific on Ubuntu
and hence need to be auto-generated to get it working on aarch64.  Use
pkg-config to do this and ignore errors because Fedora installs the
library and headers in the default paths and does not need any
additional include and lib directories.