Bump up tolerance for A-B for EqCubicSolveReal
The Test_cubiceq02 test (rightly) expects EqCubicSolveReal to return 2
real results for the equation `y(x) = x³ + x²` but this fails on
arm64.  This is because the error tolerance for A-B (i.e. the
imaginary component of the two roots) is set to 1e-14, which is too
low for arm64, which produces an error of the order of 1e-9.

Tolerate a slightly higher error value in the interest of getting a
correct result on arm64.